Refactor theme and palette management system
- Remove `AppContext` and related utility functions, consolidating theme configuration into a more robust palette-based system.
- Introduce `Palettes` and `PalettesGenerateConfig` to centralize color scheme generation logic, supporting multiple Material Design color specifications (2021, 2025, 2026).
- Refactor `AppModel` and `AppSeed` to remove dependencies on `AppContext` and file-based preferences.
- Update `AppProjector` to use `SystemPalettes` and `FractalColorsProviderByPalettes` for dynamic color rendering.
- Refactor `TonalPalette` from a class to an interface, providing a default implementation that supports HCT-based color retrieval.
- Migrate `PaletteType` and related utilities to a dedicated `palette` package.
- Simplify `AndroidViewModel` factory by removing unused context and hue parameters.
- Adjust `Tone` and `Contrast` usage to use integer-based raw values where appropriate.
- Update version code to 2 in `version.properties`.

+++ app/projector/src/commonMain/kotlin/org/hnau/commons/app/projector/fractal/utils/color/contrast/BaseWithDecayContrastExt.kt
@@ -2,14 +2,14 @@ package org.hnau.commons.app.projector.fractal.utils.color.contrast
 
 import org.hnau.commons.app.model.theme.color.Contrast
 import org.hnau.commons.app.projector.fractal.utils.BaseWithDecay
-import org.hnau.commons.app.projector.fractal.utils.double
+import org.hnau.commons.app.projector.fractal.utils.int
 import org.hnau.commons.app.projector.fractal.utils.map
 
 
 fun BaseWithDecay.Companion.contrast(
     initial: Contrast,
     decay: Double,
-): BaseWithDecay<Contrast> = BaseWithDecay.double(
+): BaseWithDecay<Contrast> = BaseWithDecay.int(
     initial = initial.contrast,
     decay = decay,
     baseline = Contrast.min.contrast
@@ -18,7 +18,7 @@ fun BaseWithDecay.Companion.contrast(
 
 
 private val containerContrast: BaseWithDecay<Contrast> = BaseWithDecay.contrast(
-    initial = Contrast(2.0),
+    initial = Contrast(2),
     decay = 0.5,
 )
 
@@ -26,7 +26,7 @@ val Contrast.Companion.container: BaseWithDecay<Contrast>
     get() = containerContrast
 
 private val contentContrast: BaseWithDecay<Contrast> = BaseWithDecay.contrast(
-    initial = Contrast(7.0),
+    initial = Contrast(7),
     decay = 0.75,
 )
